requirements.txt
pandas>=2.0.0
requests>=2.31.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
python-dotenv>=1.0.0
matplotlib>=3.8.0